services/pong/requirements.txt
flask
kafka-python
nameko
nameko-kafka
parameterized
prometheus_client
pymongo
wrapt==1.11